sync scraping to the API update schedule if possible to improve "freshness"
By using the "last update" field on the API output, and synchronizing to that plus our expected update schedule, we can record data from the API in HA as soon as possible after it's been recorded, maximising the data "freshness". If anything goes wrong with that, it'll fall back to just updating with a 5-minute period.